Use comma instead of semicolon to separate labels and values
This is easier to type, and much more readable.  The drawback is that
commas may be no longer part of a label, but it's unlikely that this is
ever desired.  For values there is even an advantage: it's easier to
understand that commas cannot be used as decimal separator, and if one
does, they get additional list elements, instead of having silently
discarded the fractional digits.