Add support for pie diagrams
Since multiple datasets don't make (much) sense for pie diagrams, we
misuse the datasets to color the individual data differently (which is
obviously important for pie diagrams), sum up the values (so a single
value per dataset is usually sufficient), and ignore the general labels
altogether.  This allows us to avoid complicating the model and the UI
even more.