Add rudimentary support for "power charts"
We'll only implement a small subset of Chart.js features, and this is
a rather opinionated way, so defining charts is understandable for non-
power users.  To make the plugin still somewhat useful for power users,
we add `chart_advanced()` which will just read a manually prepared JSON
configuration for Chart.js.  No validation or further checks, we just
pass along what the power user configured.  For now, not even any
support in the back-end.