Introduce chart transposition
While for regular (i.e. non-transposed) charts the datasets are series,
for transposed charts the first values of each dataset are a series,
the second values of each dataset are the next series, and so on.  The
labels are swapped accordingly.  Since the colors in the legend make no
sense for transposed charts, we don't display them.

Pie charts should (almost?) always be transposed (that supersedes their
prior special handling), transposition can also make sense for bar
charts (but probably not for line chart).

The concept is a bit hard to explain, so we'll add some examples later
which should make it very easy for users to grasp the concept.