Add check feature
This is mostly useful for people editing the XML files manually, but
doesn't hurt to present it to others as well – the check is fast, and
seeing a green success message is likely good.

It might also be helpful to popularize using XML with schema validation
for data storage, since this is actually much simpler than doing proper
validation on some proprietary format – and of course much better than
to not validate input at all (there be dragons).